requirements.txt
requests==2.31.0
confluent-kafka==2.3.0
pyspark==3.4.1
pymongo==4.6.0
python-dotenv==1.0.0
pandas==2.1.4
numpy==1.24.3
scikit-learn==1.3.2
great-expectations==0.18.8
pytest==7.4.3
pytest-cov==4.1.0
black==23.12.1
flake8==6.1.0
kafka-python==2.0.2
avro-python3==1.11.3
boto3==1.34.0
minio==7.2.0
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.2
redis==5.0.1
cassandra-driver==3.29.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
alembic==1.13.1
prometheus-client==0.19.0
grafana-api==1.0.3
apache-superset==3.0.3
jupyter==1.0.0
jupyterlab==4.0.10
ipykernel==6.27.1
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.17.0